requirements.txt
discord
docker
PyGithub
StackAPI
pymongo
dnspython
Pillow
